Align tiles

+++ exporter/dump_obj.js
@@ -28,6 +28,7 @@ async function run() {
 		const octName = OCTANTS.length > 3 ? `${OCTANTS.slice(0,3).join('+')}+etc` : OCTANTS.join('+');
 		const objDir = path.join(DUMP_OBJ_DIR, `${octName}-${MAX_LEVEL}-${rootEpoch}`);
 		console.log(objDir);
+		if (fs.existsSync(objDir)){ return; }
 		fs.removeSync(objDir);
 		fs.ensureDirSync(objDir);
 		objCtx = initCtxOBJ(objDir);
